@@ -75,7 +75,7 @@
   "private": true,
   "dependencies": {
     "@sentry/electron": "^2.5.4",
-    "@sentry/node": "^6.15.1",
+    "@sentry/node": "^6.16.1",
     "agentkeepalive": "^4.1.4",
     "ass-compiler": "^0.1.4",
     "async-await-parallel": "^1.0.0",
@@ -187,11 +187,11 @@
     "@types/uuid": "^8.3.3",
     "@types/which": "^2.0.1",
     "@types/yesql": "^4.1.1",
-    "@typescript-eslint/eslint-plugin": "4",
-    "@typescript-eslint/parser": "4",
+    "@typescript-eslint/eslint-plugin": "5",
+    "@typescript-eslint/parser": "5",
     "chai": "^4.3.4",
     "electron": "^16.0.5",
-    "eslint": "7",
+    "eslint": "8",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-security": "^1.4.0",
@@ -199,7 +199,7 @@
     "eslint-plugin-you-dont-need-lodash-underscore": "^6.12.0",
     "husky": "^7.0.4",
     "i18next-scanner": "^3.1.0",
-    "lint-staged": "^12.1.2",
+    "lint-staged": "^12.1.3",
     "mocha": "^9.1.3",
     "postgrator-cli": "^4.0.0",
     "prettier": "2.5.1",
