@@ -1,8 +1,9 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
-const en = require('../locales/en.json');
-const fr = require('../locales/fr.json');
-const es = require('../locales/es.json');
+
+import en from '../locales/en.json';
+import es from '../locales/es.json';
+import fr from '../locales/fr.json';
 
 i18n
 	// use react-i18next
